buildRustCrate: dont clobber src /w gen code
If a crate happened to have a build script that generates a file with
the name that’s the same with another file in `src/`, this little
snippet of code would overwrite the source with generated code…

Having looked at the history I failed to find justification for this
behaviour and since it makes actual crates (rustversion) fail to build,
this snippet has to go. See ya!

+++ pkgs/build-support/rust/build-rust-crate/configure-crate.nix
@@ -139,13 +139,8 @@ in ''
        | sed -e "s/cargo:\([^=]*\)=\(.*\)/export DEP_$(echo $CRATENAME)_\U\1\E=\2/" > target/env
 
      set -e
-     if [[ -n "$(ls target/build/${crateName}.out)" ]]; then
-
-        if [[ -e "${libPath}" ]]; then
-           cp -r target/build/${crateName}.out/* $(dirname ${libPath}) #*/
-        else
-           cp -r target/build/${crateName}.out/* src #*/
-        fi
+     if [[ -n "$(ls target/build/${crateName}.out)" && -e "${libPath}" ]]; then
+       cp -r target/build/${crateName}.out/* $(dirname ${libPath}) #*/
      fi
   fi
   runHook postConfigure
